fix repetition bug; add logging
