+++ collect_responses.py
@@ -1,10 +1,8 @@
-import os
 import json
+import logging
+
 from dotenv import load_dotenv
-from clients.gemini import GeminiClient
-from clients.deepseek import DeepSeekClient
-from clients.chatgpt import ChatGPTClient
-from clients.cloud_translation import GoogleCloudTranslationClient
+from helpers import generate_output_schema, chat_gemini, chat_chatgpt, chat_google_translate, chat_deepseek
 
 PROMPT_FILES = [
   "prompts/prompt_simple.txt",
@@ -46,152 +44,43 @@ DISASTERS = [
   #"a contagious disease called SARD-26",
 ]
 
-def chat_gemini(language, disaster, prompt):
-  gemini_client = GeminiClient(key=os.getenv("GEMINI_API_KEY"))
-  return gemini_client.chat(
-    prompt_file=prompt,
-    language=language,
-    disaster=disaster
-  )
+# Configure logging
+logger = logging.getLogger(__name__)
 
-def chat_deepseek(language, disaster, prompt):
-  deepseek_client = DeepSeekClient(key=os.getenv("OPENROUTER_API_KEY"))
-  return deepseek_client.chat(
-    prompt_file=prompt,
-    language=language,
-    disaster=disaster
-  )
-
-def chat_chatgpt(language, disaster, prompt):
-  chatgpt_client = ChatGPTClient(key=os.getenv("AZURE_OPENAI_API_KEY"), base_url=os.getenv("AZURE_OPENAI_ENDPOINT"), deployment_name=os.getenv("AZURE_DEPLOYMENT_NAME"))
-  return chatgpt_client.chat(
-    prompt_file=prompt,
-    language=language,
-    disaster=disaster
-  )
-
-def chat_google_translate(language, disaster, prompt):
-  google_translate_client = GoogleCloudTranslationClient()
-  return google_translate_client.chat(
-    prompt_file=prompt,
-    language=language,
-    disaster=disaster
-  )
+logging.basicConfig(
+    level=logging.INFO,
+    filename="responses.log",
+    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s"
+)
+logger = logging.getLogger(__name__)
 
 if __name__ == "__main__":
   load_dotenv()
 
-  prompt_breakdown_json = {
-    "prompt_simple.txt": [],
-    "prompt_persona.txt": [],
-    "prompt_one_shot.txt": [],
-    "prompt_cross_lingual_alignment.txt": [],
-    "prompt_chain_of_translation.txt": []
-  }
+  output_json = generate_output_schema()
 
-  breakdown_json = {
-    "spanish": {
-      "flood": { **prompt_breakdown_json },
-      "extreme_wind": { **prompt_breakdown_json },
-      "fire": { **prompt_breakdown_json },
-      "boil_water_notice": { **prompt_breakdown_json },
-      "911 outage": { **prompt_breakdown_json }
-    },
-    "haitian_creole": {
-      "flood": { **prompt_breakdown_json },
-      "extreme_wind": { **prompt_breakdown_json },
-      "fire": { **prompt_breakdown_json },
-      "boil_water_notice": { **prompt_breakdown_json },
-      "911_outage": { **prompt_breakdown_json }
-    },
-    "vietnamese": {
-      "flood": { **prompt_breakdown_json },
-      "extreme_wind": { **prompt_breakdown_json },
-      "fire": { **prompt_breakdown_json },
-      "boil_water_notice": { **prompt_breakdown_json },
-      "911_outage": { **prompt_breakdown_json }
-    },
-    "arabic": {
-      "flood": { **prompt_breakdown_json },
-      "extreme_wind": { **prompt_breakdown_json },
-      "fire": { **prompt_breakdown_json },
-      "boil_water_notice": { **prompt_breakdown_json },
-      "911_outage": { **prompt_breakdown_json }
-    },
-    "mandarin": {
-      "flood": { **prompt_breakdown_json },
-      "extreme_wind": { **prompt_breakdown_json },
-      "fire": { **prompt_breakdown_json },
-      "boil_water_notice": { **prompt_breakdown_json },
-      "911_outage": { **prompt_breakdown_json }
-    }
-  }
+  for language in LANGUAGES:
+    language_name = language.replace(" ", "_").lower()
 
-  output_json = {
-    "chatgpt": { **breakdown_json },
-    "gemini": { **breakdown_json },
-    "deepseek": { **breakdown_json },
-    "google_translate": {
-      "spanish": { 
-        "flood": [],
-        "extreme_wind": [],
-        "fire": [],
-        "boil_water_notice": [],
-        "911_outage": []
-      },
-      "arabic": {
-        "flood": [],
-        "extreme_wind": [],
-        "fire": [],
-        "boil_water_notice": [],
-        "911_outage": []
-      },
-      "mandarin": {
-        "flood": [],
-        "extreme_wind": [],
-        "fire": [],
-        "boil_water_notice": [],
-        "911_outage": []
-      },
-      "vietnamese": {
-        "flood": [],
-        "extreme_wind": [],
-        "fire": [],
-        "boil_water_notice": [],
-        "911_outage": []
-      },
-      "haitian_creole": {
-        "flood": [],
-        "extreme_wind": [],
-        "fire": [],
-        "boil_water_notice": [],
-        "911_outage": []
-      }
-    }
-  }
-
-  for language in ["Haitian Creole"]: # LANGUAGES
-    if language == "Haitian Creole":
-      language_name = "haitian_creole"
-    else:
-      language_name = language.lower()
-
-    for disaster in ["a fire"]: # DISASTERS
+    for disaster in DISASTERS:
       disaster_name = disaster.replace("a ", "").replace(" ", "_")
-      for prompt in ["prompts/prompt_simple.txt"]: # PROMPT_FILES
+      for prompt in PROMPT_FILES:
         prompt_name = prompt.replace("prompts/", "")
         
+        logger.info(f"Running {language_name}: {disaster_name}: {prompt_name}: Gemini")
         gemini_output = chat_gemini(language=language, disaster=disaster, prompt=prompt)
         output_json["gemini"][language_name][disaster_name][prompt_name].append(gemini_output)
 
+        logger.info(f"Running {language_name}: {disaster_name}: {prompt_name}: ChatGPT")
         chatgpt_output = chat_chatgpt(language=language, disaster=disaster, prompt=prompt)
         output_json["chatgpt"][language_name][disaster_name][prompt_name].append(chatgpt_output)
-
+        
+        logger.info(f"Running {language_name}: {disaster_name}: {prompt_name}: DeepSeek")
         deepseek_output = chat_deepseek(language=language, disaster=disaster, prompt=prompt)
         output_json["deepseek"][language_name][disaster_name][prompt_name].append(deepseek_output)
 
+      logger.info(f"Running {language_name}: {disaster_name}: Google Translate")
       prompt_file = f"prompts/{disaster_name}.txt"
-
       google_translate_output = chat_google_translate(language=language, disaster=disaster, prompt=prompt_file)
       output_json["google_translate"][language_name][disaster_name].append(google_translate_output)
